refactor(core): unify ActorPath into pitboss-core, re-export from cli
Pre-fix: two ActorPath types — pitboss-core::control_protocol::ActorPath
(transparent newtype, no constructors, no Display) and pitboss-cli::
dispatch::actor::ActorPath (same wire shape, plus new/child/leaf/Display
and pub(crate) tuple field). Adding a method to one meant remembering
to mirror it in the other; the validator on #481 noted "ActorPath has
been partially moved" but the burden was real.

Move the richer API down to pitboss-core: add new/child/leaf/Display
to the canonical type, port the Display tests inline, and add two new
inline tests for child/leaf. pitboss-cli/dispatch/actor now re-exports
pitboss_core::control_protocol::ActorPath via `pub use`, so existing
`crate::dispatch::actor::ActorPath` imports keep working byte-for-byte
across the cli, web, and tui consumers (and test sites in pitboss-web
that reach into `pitboss_cli::dispatch::actor::ActorPath`).

The typed EventEnvelope stays in pitboss-cli per the architectural rule
in pitboss-core/CLAUDE.md ("the typed ControlEvent enum carries the MCP
policy module and approval-handling types, none of which belong here").
Unifying ActorPath alone — without dragging ApprovalRule down — is the
scope the architecture allows. stream.rs's preamble comment updated to
reflect the partial unification.

Wire format is unchanged (both ActorPath types already serialized
transparently as a JSON array of strings).

Closes #481
