+++ crates/pitboss-cli/src/dispatch/actor.rs
@@ -4,11 +4,17 @@
 //! and TUI display.
 
 use serde::{Deserialize, Serialize};
-use std::fmt;
 
 /// Unique actor identifier.
 /// String is fine for depth ≤ 3; revisit if ActorPath clones show up in profiles.
 pub type ActorId = String;
+
+/// Tree-path from root to the actor that produced an event (#481).
+/// Single source of truth lives in `pitboss-core` so this crate, the
+/// unified consumer stream, and the read-side consumers share one
+/// definition. Re-exported here so existing
+/// `pitboss_cli::dispatch::actor::ActorPath` imports keep working.
+pub use pitboss_core::control_protocol::ActorPath;
 
 /// Actor role for authz in depth-2 dispatch model.
 ///
@@ -30,73 +36,9 @@ pub enum ActorRole {
     Worker,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, Default)]
-#[serde(transparent)]
-pub struct ActorPath(pub(crate) Vec<ActorId>);
-
-impl fmt::Display for ActorPath {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0.join("→"))
-    }
-}
-
-impl ActorPath {
-    /// Construct from a slice of actor ids.
-    pub fn new<I, S>(ids: I) -> Self
-    where
-        I: IntoIterator<Item = S>,
-        S: Into<ActorId>,
-    {
-        ActorPath(ids.into_iter().map(Into::into).collect())
-    }
-
-    /// Append an actor id to produce a deeper path.
-    pub fn child(&self, id: impl Into<ActorId>) -> Self {
-        let mut next = self.0.clone();
-        next.push(id.into());
-        ActorPath(next)
-    }
-
-    /// Last segment (the actor id this path identifies).
-    pub fn leaf(&self) -> Option<&ActorId> {
-        self.0.last()
-    }
-
-    /// Depth in the tree (root = 1, sub-lead = 2, worker = 3).
-    pub fn depth(&self) -> usize {
-        self.0.len()
-    }
-
-    /// Returns `true` if the path contains no segments.
-    /// Used as a `skip_serializing_if` predicate on `EventEnvelope.actor_path`
-    /// to keep the wire format backward-compatible: an empty path is simply
-    /// absent from the serialized JSON.
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn actor_path_renders_with_arrow_separator() {
-        let path = ActorPath(vec!["root".into(), "S1".into(), "W3".into()]);
-        assert_eq!(path.to_string(), "root→S1→W3");
-    }
-
-    #[test]
-    fn actor_path_renders_root_only() {
-        let path = ActorPath(vec!["root".into()]);
-        assert_eq!(path.to_string(), "root");
-    }
-
-    #[test]
-    fn actor_path_renders_empty_as_empty_string() {
-        let path = ActorPath::default();
-        assert_eq!(path.to_string(), "");
-    }
 
     #[test]
     fn actor_role_round_trips_through_serde() {
